make CarrierAgentTracker an interface and inject it. Make CarrierControlerListener the Provider<CarrierAgentTracker>
