+++ contribs/freight/src/main/java/org/matsim/contrib/freight/controler/CarrierAgent.java
@@ -9,7 +9,6 @@ import java.util.Map;
 
 import org.apache.log4j.Logger;
 import org.matsim.api.core.v01.Id;
-import org.matsim.api.core.v01.TransportMode;
 import org.matsim.api.core.v01.events.*;
 import org.matsim.api.core.v01.network.Link;
 import org.matsim.api.core.v01.population.Activity;
@@ -220,7 +219,7 @@ class CarrierAgent
 		Gbl.assertNotNull(carrierScoringFunction);
 	}
 
-	public CarrierAgent( CarrierAgentTracker lspCarrierTracker, Carrier carrier ){
+	public CarrierAgent(CarrierAgentTracker lspCarrierTracker, Carrier carrier ){
 		lspTracker = lspCarrierTracker;
 		this.carrier = carrier;
 		this.id = carrier.getId();
